pyproject.toml
[tool.poetry]
name = "ai-category-extractor"
version = "0.1.0"
description = "AI-powered category extraction for e-commerce sites"
authors = ["Ashley Coleman <ashley@example.com>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
strands-agents = "^0.3.0"
playwright = "^1.40.0"
asyncpg = "^0.29.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
click = "^8.1.0"
rich = "^13.7.0"
loguru = "^0.7.2"
anthropic = "^0.8.0"
openai = "^1.0.0"
httpx = "^0.25.0"
tenacity = "^8.2.0"
beautifulsoup4 = "^4.12.0"
lxml = "^5.0.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.1.0"
black = "^23.12.0"
mypy = "^1.8.0"
ruff = "^0.1.0"
types-beautifulsoup4 = "^4.12.0.7"

[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "-ra --strict-markers"
markers = [
  "e2e: marks end-to-end tests hitting external sites",
  "slow: marks long-running tests"
]
